Bug #20: code refactoring
